some more tweaks for bug 30311. Set some (unused) bits in  FontFormatting to match Excel.
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@646194 13f79535-47bb-0310-9956-ffa450edef68

+++ src/testcases/org/apache/poi/hssf/record/TestCFRuleRecord.java
@@ -230,7 +230,7 @@ public final class TestCFRuleRecord extends TestCase
 		fontFormatting.setFontColorIndex((short)10);
 		assertEquals(10,fontFormatting.getFontColorIndex());
 
-		fontFormatting.setFontHeight((short)100);
+		fontFormatting.setFontHeight(100);
 		assertEquals(100,fontFormatting.getFontHeight());
 
 		fontFormatting.setFontOutlineModified(false);
